+++ backend/src/main/java/org/catalogueoflife/editor/name/NameUsageController.java
@@ -31,11 +31,11 @@ public class NameUsageController {
   }
 
   @GetMapping
-  public List<NameUsageResponse> list(@PathVariable long pid,
+  public List<NameUsageResponse> list(@PathVariable int pid,
       @RequestParam(required = false) String q,
       @RequestParam(defaultValue = "50") int limit,
       @RequestParam(defaultValue = "0") int offset) {
-    long uid = currentUser.require().getId();
+    int uid = currentUser.require().getId();
     return (q == null || q.isBlank())
         ? service.list(uid, pid, limit, offset)
         : service.search(uid, pid, q, limit, offset);
@@ -43,42 +43,42 @@ public class NameUsageController {
 
   @PostMapping
   @ResponseStatus(HttpStatus.CREATED)
-  public NameUsageResponse create(@PathVariable long pid, @Valid @RequestBody CreateNameUsageRequest req) {
-    long uid = currentUser.require().getId();
+  public NameUsageResponse create(@PathVariable int pid, @Valid @RequestBody CreateNameUsageRequest req) {
+    int uid = currentUser.require().getId();
     return service.create(uid, pid, req);
   }
 
   @GetMapping("/{id}")
-  public NameUsageResponse get(@PathVariable long pid, @PathVariable long id) {
-    long uid = currentUser.require().getId();
+  public NameUsageResponse get(@PathVariable int pid, @PathVariable int id) {
+    int uid = currentUser.require().getId();
     return service.get(uid, pid, id);
   }
 
   @PutMapping("/{id}")
-  public NameUsageResponse update(@PathVariable long pid, @PathVariable long id,
+  public NameUsageResponse update(@PathVariable int pid, @PathVariable int id,
       @Valid @RequestBody UpdateNameUsageRequest req) {
-    long uid = currentUser.require().getId();
+    int uid = currentUser.require().getId();
     return service.update(uid, pid, id, req);
   }
 
   @DeleteMapping("/{id}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void delete(@PathVariable long pid, @PathVariable long id) {
-    long uid = currentUser.require().getId();
+  public void delete(@PathVariable int pid, @PathVariable int id) {
+    int uid = currentUser.require().getId();
     service.delete(uid, pid, id);
   }
 
   @PutMapping("/{id}/synonym-of/{acceptedId}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void linkSynonym(@PathVariable long pid, @PathVariable long id, @PathVariable long acceptedId) {
-    long uid = currentUser.require().getId();
+  public void linkSynonym(@PathVariable int pid, @PathVariable int id, @PathVariable int acceptedId) {
+    int uid = currentUser.require().getId();
     service.linkSynonym(uid, pid, id, acceptedId);
   }
 
   @DeleteMapping("/{id}/synonym-of/{acceptedId}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void unlinkSynonym(@PathVariable long pid, @PathVariable long id, @PathVariable long acceptedId) {
-    long uid = currentUser.require().getId();
+  public void unlinkSynonym(@PathVariable int pid, @PathVariable int id, @PathVariable int acceptedId) {
+    int uid = currentUser.require().getId();
     service.unlinkSynonym(uid, pid, id, acceptedId);
   }
 }
